Move SingleTests
