+++ Tests/Batch1/SimpleTypes/SingleTests.cs
@@ -1,5 +1,6 @@
 ﻿using Bridge.Test.NUnit;
 using System;
+using System.Globalization;
 
 namespace Bridge.ClientTest.SimpleTypes
 {
@@ -8,14 +9,25 @@ namespace Bridge.ClientTest.SimpleTypes
     public class SingleTests
     {
         [Test]
-        public void TypePropertiesAreCorrect()
+        public void TypePropertiesAreCorrect_SPI_1717()
         {
             Assert.True((object)(float)0.5 is float);
             Assert.AreEqual("System.Single", typeof(float).FullName);
-
+            Assert.False(typeof(float).IsClass);
+            Assert.True(typeof(IComparable<float>).IsAssignableFrom(typeof(float)));
+            Assert.True(typeof(IEquatable<float>).IsAssignableFrom(typeof(float)));
+            Assert.True(typeof(IFormattable).IsAssignableFrom(typeof(float)));
             object f = (float)0;
             Assert.True(f is float);
+            Assert.True(f is IComparable<float>);
+            Assert.True(f is IEquatable<float>);
             Assert.True(f is IFormattable);
+
+            var interfaces = typeof(float).GetInterfaces();
+            Assert.AreEqual(4, interfaces.Length);
+            Assert.True(interfaces.Contains(typeof(IComparable<float>)));
+            Assert.True(interfaces.Contains(typeof(IEquatable<float>)));
+            Assert.True(interfaces.Contains(typeof(IFormattable)));
         }
 
         private T GetDefaultValue<T>()
@@ -32,7 +44,7 @@ namespace Bridge.ClientTest.SimpleTypes
         [Test]
         public void CreatingInstanceReturnsZero()
         {
-            Assert.AreEqual(0, Activator.CreateInstance<float>());
+            Assert.AreStrictEqual(0, Activator.CreateInstance<float>());
         }
 
         [Test]
@@ -41,7 +53,7 @@ namespace Bridge.ClientTest.SimpleTypes
             double zero = 0;
             Assert.True((float)(object)float.MinValue < -3.4e38 && (float)(object)float.MinValue > -3.5e38, "MinValue should be correct");
             Assert.True((float)(object)float.MaxValue > 3.4e38 && (float)(object)float.MaxValue < 3.5e38, "MaxValue should be correct");
-            Assert.AreEqual(1.401298e-45, /*1.40129846432482E-45*/float.Epsilon, "Epsilon should be correct");
+            Assert.AreEqual(1.401298e-45, float.Epsilon, "Epsilon should be correct");
             Assert.True(float.IsNaN(float.NaN), "NaN should be correct");
             Assert.AreStrictEqual(1 / zero, float.PositiveInfinity, "PositiveInfinity should be correct");
             Assert.AreStrictEqual(-1 / zero, float.NegativeInfinity, "NegativeInfinity should be correct");
@@ -60,10 +72,29 @@ namespace Bridge.ClientTest.SimpleTypes
         }
 
         [Test]
-        public void IFormattableToStringWorks()
+        public void ToStringWithFormatWorks()
         {
             Assert.AreEqual("123", ((float)291.0).ToString("x"));
         }
+
+        [Test]
+        public void ToStringWithFormatAndProviderWorks()
+        {
+            Assert.AreEqual("123", ((float)291.0).ToString("x", CultureInfo.InvariantCulture));
+        }
+
+        [Test]
+        public void IFormattableToStringWorks()
+        {
+            Assert.AreEqual("123", ((IFormattable)((float)291.0)).ToString("x", CultureInfo.InvariantCulture));
+        }
+
+        // Not C# API
+        //[Test]
+        //public void LocaleFormatWorks()
+        //{
+        //    Assert.AreEqual(((float)291.0).LocaleFormat("x"), "123");
+        //}
 
         [Test]
         public void ToStringWorks()
@@ -108,11 +139,12 @@ namespace Bridge.ClientTest.SimpleTypes
         }
 
         [Test]
-        public void IsPositiveInfinityWorks()
+        public void IsPositiveInfinityWorks_SPI_1600()
         {
             float inf = 1.0f / 0.0f;
-            // TODO Bug
-            //Assert.True (float.IsPositiveInfinity(inf));
+            // #1600
+            Assert.True(float.IsPositiveInfinity(inf));
+
             Assert.False(float.IsPositiveInfinity(-inf));
             Assert.False(float.IsPositiveInfinity(0.0f));
             Assert.False(float.IsPositiveInfinity(Single.NaN));
